requirements.txt
django
pillow
django-allauth
python-decouple